HIS-75: move custom node styling from flow to treenode
