+++ histree-frontend/src/components/TreeNode.scss
@@ -1,25 +1,25 @@
-.treenodecard {
-	margin: 0 1em 0 1em;
-	width: 100px;
-	height: 25px;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-}
-
-.treenodechild {
-	white-space: nowrap;
-	overflow: hidden;
-	text-overflow: ellipsis;
-	min-width: 0;
-}
-
-.react-flow__node-dataNode {
+.tree-node {
+	border-radius: 5px;
 	font-size: 12px;
 	background: white;
 	border: 1px solid #555;
-	border-radius: 5px;
 	text-align: center;
+
+	&-card {
+		margin: 0 1em 0 1em;
+		width: 100px;
+		height: 25px;
+		display: flex;
+		justify-content: center;
+		align-items: center;
+	}
+
+	&-child {
+		white-space: nowrap;
+		overflow: hidden;
+		text-overflow: ellipsis;
+		min-width: 0;
+	}
 }
 
 .handle_loading {
