+++ histree-frontend/src/components/TreeNode.tsx
@@ -1,25 +1,22 @@
-import React, { useEffect, useRef } from 'react';
+import React, { CSSProperties, useEffect, useRef } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import {
 	setSelected,
 	getRenderContent,
 	getNodeLookup,
 	setNodeLookup,
-	AppDispatch,
 	setNodeLookupDown,
 	setNodeLookupUp,
 	getRenderMode,
 	setComparisonNode,
 	setEdgeInfo,
 	getCompareNodes,
-	setRelationship,
-	setNodeOnScreen
+	setNodeOnScreen,
+	getEdgeInfo
 } from '../stores/base';
-import { AdjList, HandleStatus, NodeInfo, RenderContent } from '../models';
+import { HandleStatus, NodeId, NodeInfo } from '../models';
 import './TreeNode.scss';
 import { Handle, Position } from 'reactflow';
-import { DataSuccess, fetchRelationship, fetchSelectedExpansion } from '../services';
-import { cleanseBranches, findPathBetweenTwoNodes } from '../utils/utils';
 import { useIsInViewport } from '../utils/viewport';
 
 const nodeClassMap: Record<HandleStatus, string> = {
@@ -38,6 +35,7 @@ const TreeNode = ({ data }: { data: NodeInfo }) => {
 	const comparisonNodes = useSelector(getCompareNodes);
 	const renderContent = useSelector(getRenderContent);
 	const nodeLookup = useSelector(getNodeLookup);
+	const edgeInfo = useSelector(getEdgeInfo);
 
 	useEffect(() => {
 		if (isInView) {
@@ -121,8 +119,17 @@ const TreeNode = ({ data }: { data: NodeInfo }) => {
 		}
 	};
 
+	const getNodeStyle = (nodeid: NodeId): CSSProperties => {
+		if (comparisonNodes.first && comparisonNodes.first.id === nodeid ||
+			comparisonNodes.second && comparisonNodes.second.id === nodeid ||
+			edgeInfo[nodeid] !== undefined) {
+			return { color: 'orange', borderColor: 'orange' };
+		}
+		return {}
+	}
+
 	return (
-		<div ref={ref}>
+		<div ref={ref} className="tree-node" style={getNodeStyle(data.id)}>
 			{nodeLookup[data.id].visible ? <>
 				<Handle
 					className={nodeClassMap[nodeLookup[data.id].upExpanded]}
@@ -131,8 +138,8 @@ const TreeNode = ({ data }: { data: NodeInfo }) => {
 					isConnectable
 					onClick={handleExpandParents}
 				/>
-				<div className="treenodecard" onClick={handleNodeClick}>
-					<div className="treenodechild">{data.name}</div>
+				<div className="tree-node-card" onClick={handleNodeClick}>
+					<div className="tree-node-child">{data.name}</div>
 				</div>
 
 				<Handle
